+++ echo/ws-echo/js-client/addr.js
@@ -1,2 +1,3 @@
     // var serverAddress = '192.168.8.3:10000';
     var serverAddress = '127.0.0.1:10000';
+    // var serverAddress = 'localhost:10000';